+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.0 - 2020-08-17
+
+### Added
+
+- `[NXMClientConfig AMS]` static method.
+
+### Fixed
+
+- Custom events parsing.
+
 ## 2.2.2 - 2020-07-20
 
 ### Fixed
@@ -489,4 +499,3 @@ NXMClient myClient = ...;
 
 - Please refer to list of features and usage: <https://developer.nexmo.com/>
 - **Cocoapods**: <https://cocoapods.org/pods/nexmoclient>
-